Time: 7 ms (76.99%) | Memory: 20.3 MB (49.64%) - LeetSync

+++ Trees/98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -12,10 +12,10 @@
 class Solution {
 public:
     bool isValidBST(TreeNode* root) {
-        return solve(root, LLONG_MIN, LLONG_MAX);
+        return solve(root, LONG_MIN, LONG_MAX);
     }
 
-    bool solve(TreeNode* root,long long mini, long long maxi) {
+    bool solve(TreeNode* root,long mini, long maxi) {
         if(root == NULL) return true;
         if(root->val <= mini || root->val >= maxi) return false;
         return solve(root->left,mini, root->val) && solve(root->right,root->val,maxi);
